housekeeping refactoring navigation and components folder structure
