+++ frontend/src/interfaces/overviewCardInterface.ts
