+++ frontend/src/interfaces/socialInterface.ts
