+++ frontend/src/interfaces/transactionHistoryInterface.ts
