+++ frontend/src/state/index.ts
@@ -1,4 +1,5 @@
+
+// Export store and reducers
 export * as actionCreators from "./action-creators";
 export * from "./reducers/index";
 export * from './store';
-// Export redux files
